Complete 6-phase AI orchestration system with first successful demo
Major Features:
- AI-powered planning in WebScraperRouter (no hardcoded patterns)
- 6-phase flow: Intent → CEO → Department Planning → Execution → Synthesis → Personality
- SystemIntelRouter → AndroidIntelRouter rename for Android specificity
- Personality survival after tool execution (Phase 5 implementation)
- UI improvements: No/Edit/Yes buttons, plan validation, phase isolation
- Historic first successful orchestration demo video and GIF

Technical Improvements:
- WebScraperRouter uses AI models to analyze requests and create intelligent plans
- CEO role simplified to pure department selection (no planning/synthesis)
- Department routers do all AI reasoning and execution planning
- Synthesis role clarified as data amalgamation only
- PersonalityResponsePhase added to UI components
- Button text optimized for mobile layout
- Phase thinking isolation prevents UI collision

Demo Integration:
- Added screen recording of complete "What's on Hacker News?" flow
- GIF auto-plays for immediate impact (20.1MB, 30s)
- Full video available in expandable section (45.8MB, 1:34)
- Showcases real-time AI reasoning across all 6 phases

Architecture: Pure AI reasoning throughout - routers analyze requests intelligently rather than using hardcoded if/else patterns. This enables handling any web request: "What's on Reddit?", "Check CNN headlines", "GitHub trending" etc.
