+++ app/src/main/java/com/yourname/coquettemobile/core/models/ChatMessage.kt
@@ -47,6 +47,12 @@ sealed class OrchestrationPhase {
         override val thinking: List<String>,
         val response: String
     ) : OrchestrationPhase()
+
+    @Serializable
+    data class PersonalityResponsePhase(
+        override val thinking: List<String>,
+        val response: String
+    ) : OrchestrationPhase()
 }
