+++ app/src/main/java/com/yourname/coquettemobile/core/orchestration/routers/DesktopExploitRouter.kt
@@ -467,7 +467,7 @@ class DesktopExploitRouter @Inject constructor(
     
     private fun extractTargetOS(context: OperationContext): String {
         // Extract OS information from previous SystemIntel results
-        val systemIntelResults = context.getResultsFromDomain(RouterDomain.SYSTEM_INTELLIGENCE)
+        val systemIntelResults = context.getResultsFromDomain(RouterDomain.ANDROID_INTELLIGENCE)
         
         for (result in systemIntelResults) {
             if (result.data.containsKey("device_info")) {
