--- app/src/main/java/com/yourname/coquettemobile/core/orchestration/routers/SystemIntelRouter.kt
@@ -1,483 +1,0 @@
-package com.yourname.coquettemobile.core.orchestration.routers
-
-import com.yourname.coquettemobile.core.logging.CoquetteLogger
-import com.yourname.coquettemobile.core.orchestration.*
-import com.yourname.coquettemobile.core.tools.*
-import javax.inject.Inject
-import javax.inject.Singleton
-
-/**
- * SystemIntelRouter - Specialized router for information gathering and reconnaissance
- * Domain expert in environment discovery, target analysis, and capability assessment
- */
-@Singleton
-class SystemIntelRouter @Inject constructor(
-    private val deviceContextTool: DeviceContextTool,
-    private val webFetchTool: WebFetchTool,
-    private val extractorTool: ExtractorTool,
-    private val summarizerTool: SummarizerTool,
-    // Note: NetworkScanTool will be implemented in future iteration
-    private val fileTool: FileTool,
-    private val grepTool: GrepTool,
-    private val globTool: GlobTool,
-    private val logger: CoquetteLogger
-) : ToolRouter {
-    
-    override val domain = RouterDomain.SYSTEM_INTELLIGENCE
-    override val name = "SystemIntelRouter"
-    
-    override val managedTools: List<MobileTool> = listOf(
-        deviceContextTool,
-        webFetchTool,
-        extractorTool,
-        summarizerTool,
-        fileTool,
-        grepTool,
-        globTool
-    )
-    
-    override val capabilities = listOf(
-        "system analysis", "environment discovery", "target reconnaissance",
-        "device information", "network scanning", "vulnerability assessment",
-        "file system analysis", "configuration discovery", "capability assessment",
-        "system profiling", "resource enumeration", "security analysis",
-        "web content fetching", "data extraction", "content summarization",
-        "web intelligence gathering", "external data retrieval"
-    )
-    
-    override val priority = 80 // High priority for intelligence gathering
-    
-    override suspend fun canHandle(operation: OperationStep): Boolean {
-        return when (operation.type) {
-            StepType.ENVIRONMENT_DISCOVERY,
-            StepType.TARGET_ANALYSIS,
-            StepType.CAPABILITY_ASSESSMENT,
-            StepType.VULNERABILITY_SCAN,
-            StepType.FILE_ENUMERATION,
-            StepType.DIRECTORY_ANALYSIS,
-            StepType.WEB_CONTENT_FETCH,
-            StepType.CONTENT_EXTRACTION,
-            StepType.DATA_SUMMARIZATION -> true
-            else -> operation.domain == domain
-        }
-    }
-    
-    override suspend fun executeStep(step: OperationStep, context: OperationContext): StepResult {
-        val startTime = System.currentTimeMillis()
-        logger.d(name, "Executing ${step.type}: ${step.description}")
-        
-        return try {
-            when (step.type) {
-                StepType.ENVIRONMENT_DISCOVERY -> discoverEnvironment(step, context)
-                StepType.TARGET_ANALYSIS -> analyzeTarget(step, context)
-                StepType.CAPABILITY_ASSESSMENT -> assessCapabilities(step, context)
-                StepType.VULNERABILITY_SCAN -> scanVulnerabilities(step, context)
-                StepType.FILE_ENUMERATION -> enumerateFiles(step, context)
-                StepType.DIRECTORY_ANALYSIS -> analyzeDirectories(step, context)
-                StepType.WEB_CONTENT_FETCH -> fetchWebContent(step, context)
-                StepType.CONTENT_EXTRACTION -> extractContent(step, context)
-                StepType.DATA_SUMMARIZATION -> summarizeData(step, context)
-                else -> StepResult.unsupported(step.id, domain)
-            }
-        } catch (e: Exception) {
-            val executionTime = System.currentTimeMillis() - startTime
-            logger.e(name, "Step execution failed: ${e.message}")
-            
-            StepResult.failure(
-                step.id,
-                domain,
-                "SystemIntel execution error: ${e.message}",
-                executionTime
-            )
-        }
-    }
-    
-    override suspend fun planSubSteps(goal: String, context: OperationContext): List<OperationStep> {
-        val subSteps = mutableListOf<OperationStep>()
-        val goalLower = goal.lowercase()
-        
-        // Always start with environment discovery
-        subSteps.add(
-            OperationStep(
-                id = "intel_env_discovery",
-                type = StepType.ENVIRONMENT_DISCOVERY,
-                domain = domain,
-                description = "Discover system environment and basic information",
-                estimatedDurationMs = 30000L
-            )
-        )
-        
-        // Add specific analysis based on goal
-        when {
-            goalLower.contains("network") -> {
-                subSteps.add(
-                    OperationStep(
-                        id = "intel_network_scan",
-                        type = StepType.TARGET_ANALYSIS,
-                        domain = domain,
-                        description = "Analyze network configuration and connectivity",
-                        dependencies = listOf("intel_env_discovery"),
-                        estimatedDurationMs = 45000L
-                    )
-                )
-            }
-            
-            goalLower.contains("file") || goalLower.contains("data") -> {
-                subSteps.add(
-                    OperationStep(
-                        id = "intel_file_analysis",
-                        type = StepType.FILE_ENUMERATION,
-                        domain = domain,
-                        description = "Enumerate and analyze file system structure",
-                        dependencies = listOf("intel_env_discovery"),
-                        estimatedDurationMs = 60000L
-                    )
-                )
-            }
-            
-            goalLower.contains("security") || goalLower.contains("vulnerability") -> {
-                subSteps.add(
-                    OperationStep(
-                        id = "intel_security_scan",
-                        type = StepType.VULNERABILITY_SCAN,
-                        domain = domain,
-                        description = "Assess security posture and potential vulnerabilities",
-                        dependencies = listOf("intel_env_discovery"),
-                        estimatedDurationMs = 90000L
-                    )
-                )
-            }
-        }
-        
-        // Always end with capability assessment
-        subSteps.add(
-            OperationStep(
-                id = "intel_capability_assessment",
-                type = StepType.CAPABILITY_ASSESSMENT,
-                domain = domain,
-                description = "Assess available capabilities for further operations",
-                dependencies = subSteps.map { it.id },
-                estimatedDurationMs = 20000L
-            )
-        )
-        
-        return subSteps
-    }
-    
-    override suspend fun validateCapabilities(operation: OperationStep): ValidationResult {
-        val requirements = mutableListOf<String>()
-        
-        when (operation.type) {
-            StepType.ENVIRONMENT_DISCOVERY -> {
-                if (!hasDeviceContextCapability()) {
-                    requirements.add("Device context access required")
-                }
-            }
-            StepType.FILE_ENUMERATION, StepType.DIRECTORY_ANALYSIS -> {
-                if (!hasFileSystemCapability()) {
-                    requirements.add("File system access required")
-                }
-            }
-            StepType.VULNERABILITY_SCAN -> {
-                if (!hasSecurityAnalysisCapability()) {
-                    requirements.add("Security analysis tools required")
-                }
-            }
-            else -> { /* No specific requirements */ }
-        }
-        
-        return if (requirements.isEmpty()) {
-            ValidationResult.valid()
-        } else {
-            ValidationResult.invalid(requirements)
-        }
-    }
-    
-    override suspend fun getInsights(context: OperationContext): RouterInsights {
-        return RouterInsights(
-            domain = domain,
-            availableCapabilities = capabilities,
-            currentConstraints = getCurrentConstraints(context),
-            recommendedApproach = getRecommendedApproach(context),
-            riskAssessment = com.yourname.coquettemobile.core.tools.RiskLevel.LOW // Intelligence gathering is typically low risk
-        )
-    }
-    
-    // Implementation methods
-    
-    private suspend fun discoverEnvironment(step: OperationStep, context: OperationContext): StepResult {
-        val startTime = System.currentTimeMillis()
-        val discoveredData = mutableMapOf<String, Any>()
-        
-        // Get device context information
-        val deviceResult = deviceContextTool.execute(mapOf("action" to "system"))
-        if (deviceResult is ToolResult.Success) {
-            discoveredData["device_info"] = deviceResult.output
-        }
-        
-        // Get battery information
-        val batteryResult = deviceContextTool.execute(mapOf("action" to "battery"))
-        if (batteryResult is ToolResult.Success) {
-            discoveredData["battery_info"] = batteryResult.output
-        }
-        
-        // Get network information
-        val networkResult = deviceContextTool.execute(mapOf("action" to "network"))
-        if (networkResult is ToolResult.Success) {
-            discoveredData["network_info"] = networkResult.output
-        }
-        
-        val executionTime = System.currentTimeMillis() - startTime
-        
-        return StepResult.success(
-            step.id,
-            domain,
-            data = discoveredData.mapValues { it.value.toString() },
-            executionTimeMs = executionTime
-        )
-    }
-    
-    private suspend fun analyzeTarget(step: OperationStep, context: OperationContext): StepResult {
-        val startTime = System.currentTimeMillis()
-        val analysisData = mutableMapOf<String, Any>()
-        
-        // Analyze previous environment discovery results
-        val envResults = context.getResultsFromDomain(domain)
-            .firstOrNull { it.stepId.contains("env_discovery") }
-        
-        if (envResults != null) {
-            analysisData["environment_analysis"] = analyzeEnvironmentData(envResults.data)
-        }
-        
-        // Add target-specific analysis based on parameters
-        val targetType = step.parameters["target_type"] as? String
-        if (targetType != null) {
-            analysisData["target_specific_analysis"] = analyzeTargetType(targetType)
-        }
-        
-        val executionTime = System.currentTimeMillis() - startTime
-        
-        return StepResult.success(
-            step.id,
-            domain,
-            data = analysisData.mapValues { it.value.toString() },
-            executionTimeMs = executionTime
-        )
-    }
-    
-    private suspend fun assessCapabilities(step: OperationStep, context: OperationContext): StepResult {
-        val startTime = System.currentTimeMillis()
-        
-        val capabilities = mutableMapOf<String, Any>()
-        capabilities["available_tools"] = managedTools.map { it.name }
-        capabilities["intelligence_capabilities"] = this.capabilities
-        capabilities["operational_readiness"] = assessOperationalReadiness(context)
-        
-        val executionTime = System.currentTimeMillis() - startTime
-        
-        return StepResult.success(
-            step.id,
-            domain,
-            data = capabilities.mapValues { it.value.toString() },
-            executionTimeMs = executionTime
-        )
-    }
-    
-    private suspend fun scanVulnerabilities(step: OperationStep, context: OperationContext): StepResult {
-        val startTime = System.currentTimeMillis()
-        
-        // This is a placeholder - actual vulnerability scanning would be more complex
-        val vulnerabilities = mutableMapOf<String, Any>()
-        vulnerabilities["scan_type"] = "basic_assessment"
-        vulnerabilities["findings"] = "Vulnerability scanning capabilities limited in current implementation"
-        vulnerabilities["recommendations"] = listOf(
-            "Implement network vulnerability scanner",
-            "Add system configuration analyzer",
-            "Include security baseline checker"
-        )
-        
-        val executionTime = System.currentTimeMillis() - startTime
-        
-        return StepResult.success(
-            step.id,
-            domain,
-            data = vulnerabilities.mapValues { it.value.toString() },
-            executionTimeMs = executionTime
-        )
-    }
-    
-    private suspend fun enumerateFiles(step: OperationStep, context: OperationContext): StepResult {
-        val startTime = System.currentTimeMillis()
-        val fileData = mutableMapOf<String, Any>()
-        
-        val searchPath = step.parameters["path"] as? String ?: "/storage/emulated/0"
-        val pattern = step.parameters["pattern"] as? String ?: "*"
-        
-        // Use GlobTool to find files matching pattern
-        val globResult = globTool.execute(mapOf(
-            "pattern" to pattern,
-            "path" to searchPath
-        ))
-        
-        if (globResult is ToolResult.Success) {
-            fileData["enumerated_files"] = globResult.output
-        }
-        
-        val executionTime = System.currentTimeMillis() - startTime
-        
-        return StepResult.success(
-            step.id,
-            domain,
-            data = fileData.mapValues { it.value.toString() },
-            executionTimeMs = executionTime
-        )
-    }
-    
-    private suspend fun analyzeDirectories(step: OperationStep, context: OperationContext): StepResult {
-        val startTime = System.currentTimeMillis()
-        val dirData = mutableMapOf<String, Any>()
-        
-        // Analyze directory structure and contents
-        dirData["analysis_type"] = "directory_structure"
-        dirData["analyzed_paths"] = step.parameters["paths"] ?: listOf("/storage/emulated/0")
-        
-        val executionTime = System.currentTimeMillis() - startTime
-        
-        return StepResult.success(
-            step.id,
-            domain,
-            data = dirData.mapValues { it.value.toString() },
-            executionTimeMs = executionTime
-        )
-    }
-    
-    private suspend fun fetchWebContent(step: OperationStep, context: OperationContext): StepResult {
-        val startTime = System.currentTimeMillis()
-        
-        val url = step.parameters["url"] as? String 
-            ?: return StepResult.failure(step.id, domain, "No URL provided for web content fetch", 0L)
-        
-        val webResult = webFetchTool.execute(mapOf("url" to url))
-        
-        val executionTime = System.currentTimeMillis() - startTime
-        
-        return if (webResult is ToolResult.Success) {
-            StepResult.success(
-                step.id,
-                domain,
-                data = mapOf(
-                    "url" to url,
-                    "content" to webResult.output,
-                    "content_length" to webResult.output.length.toString()
-                ),
-                executionTimeMs = executionTime
-            )
-        } else {
-            StepResult.failure(step.id, domain, "Web fetch failed: ${webResult.output}", executionTime)
-        }
-    }
-    
-    private suspend fun extractContent(step: OperationStep, context: OperationContext): StepResult {
-        val startTime = System.currentTimeMillis()
-        
-        val content = step.parameters["content"] as? String 
-            ?: return StepResult.failure(step.id, domain, "No content provided for extraction", 0L)
-        
-        val extractResult = extractorTool.execute(mapOf("content" to content))
-        
-        val executionTime = System.currentTimeMillis() - startTime
-        
-        return if (extractResult is ToolResult.Success) {
-            StepResult.success(
-                step.id,
-                domain,
-                data = mapOf(
-                    "extracted_content" to extractResult.output,
-                    "original_length" to content.length.toString(),
-                    "extracted_length" to extractResult.output.length.toString()
-                ),
-                executionTimeMs = executionTime
-            )
-        } else {
-            StepResult.failure(step.id, domain, "Content extraction failed: ${extractResult.output}", executionTime)
-        }
-    }
-    
-    private suspend fun summarizeData(step: OperationStep, context: OperationContext): StepResult {
-        val startTime = System.currentTimeMillis()
-        
-        val content = step.parameters["content"] as? String 
-            ?: return StepResult.failure(step.id, domain, "No content provided for summarization", 0L)
-        
-        val summaryResult = summarizerTool.execute(mapOf(
-            "content" to content,
-            "max_length" to (step.parameters["max_length"] ?: "500")
-        ))
-        
-        val executionTime = System.currentTimeMillis() - startTime
-        
-        return if (summaryResult is ToolResult.Success) {
-            StepResult.success(
-                step.id,
-                domain,
-                data = mapOf(
-                    "summary" to summaryResult.output,
-                    "original_length" to content.length.toString(),
-                    "compression_ratio" to (content.length.toFloat() / summaryResult.output.length).toString()
-                ),
-                executionTimeMs = executionTime
-            )
-        } else {
-            StepResult.failure(step.id, domain, "Content summarization failed: ${summaryResult.output}", executionTime)
-        }
-    }
-    
-    // Helper methods
-    
-    private fun hasDeviceContextCapability(): Boolean = true // DeviceContextTool is always available
-    
-    private fun hasFileSystemCapability(): Boolean = true // FileTool, GrepTool, GlobTool available
-    
-    private fun hasSecurityAnalysisCapability(): Boolean = false // Not fully implemented yet
-    
-    private fun getCurrentConstraints(context: OperationContext): List<String> {
-        val constraints = mutableListOf<String>()
-        
-        if (context.securityLevel == SecurityLevel.MINIMAL) {
-            constraints.add("Limited to basic system information only")
-        }
-        
-        return constraints
-    }
-    
-    private fun getRecommendedApproach(context: OperationContext): String {
-        return "Systematic intelligence gathering: environment → analysis → capabilities"
-    }
-    
-    private fun analyzeEnvironmentData(envData: Map<String, Any>): Map<String, Any> {
-        val analysis = mutableMapOf<String, Any>()
-        
-        analysis["data_quality"] = "good"
-        analysis["completeness"] = envData.size
-        analysis["key_findings"] = "Environment data collected successfully"
-        
-        return analysis
-    }
-    
-    private fun analyzeTargetType(targetType: String): Map<String, Any> {
-        return mapOf(
-            "target_type" to targetType,
-            "analysis_approach" to "generic_analysis",
-            "confidence" to "medium"
-        )
-    }
-    
-    private fun assessOperationalReadiness(context: OperationContext): String {
-        return when {
-            context.securityLevel == SecurityLevel.MAXIMUM -> "fully_operational"
-            context.securityLevel == SecurityLevel.ELEVATED -> "elevated_operational"
-            context.securityLevel == SecurityLevel.STANDARD -> "standard_operational"
-            else -> "limited_operational"
-        }
-    }
-}
