+++ app/src/main/java/com/yourname/coquettemobile/ui/components/orchestration/LiveOperationProgress.kt
@@ -336,7 +336,7 @@ private fun StepProgressItem(
 // Utility functions (reuse from ExecutionPlanPreviewCard)
 private fun getRouterColor(domain: RouterDomain): Color {
     return when (domain) {
-        RouterDomain.SYSTEM_INTELLIGENCE -> Color(0xFF2196F3)
+        RouterDomain.ANDROID_INTELLIGENCE -> Color(0xFF2196F3)
         RouterDomain.DESKTOP_EXPLOIT -> Color(0xFFFF5722)  
         RouterDomain.NETWORK_OPERATIONS -> Color(0xFF4CAF50)
         RouterDomain.FILE_OPERATIONS -> Color(0xFFFF9800)
@@ -350,7 +350,7 @@ private fun getRouterColor(domain: RouterDomain): Color {
 
 private fun getRouterIcon(domain: RouterDomain): String {
     return when (domain) {
-        RouterDomain.SYSTEM_INTELLIGENCE -> "🔍"
+        RouterDomain.ANDROID_INTELLIGENCE -> "🔍"
         RouterDomain.DESKTOP_EXPLOIT -> "🖱️"
         RouterDomain.NETWORK_OPERATIONS -> "🌐" 
         RouterDomain.FILE_OPERATIONS -> "📁"
